pyproject.toml
[tool.poetry]
name = "planai"
version = "0.1.0"
description = "A simple framework for coordinating classicial compute and LLM-based tasks."
authors = ["Niels Provos <provos@gmail.com>"]
readme = "README.md"
packages = [
    { include = "planai", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.8.2"
langchain = "^0.2.14"
python-dotenv = "^1.0.1"
ollama = "^0.3.1"
diskcache = "^5.6.3"
paramiko = "^3.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
